feat: Add the third step of the post a listing flow
Also, the AnimatedSwitcher widget has been used in post_listing_screen instead
of the Stack widget.

Note: In this commit, a model called TempGear has been used. It's a temporary
model as a replacement for the Gear model. The response of posting a listing
has to be changed form the backend in order to reuse the Gear model again.