feat: Add images tab view
User can now take a photo by camera or pick from the gallery.
The picked photo is uploaded but it's not saved with the listing yet.
User also can delete the picked photos (locally only since they are not saved with the listing).
User can preview uploaded photos in a dialog.