feat: Get the total price in listing details screen feat: Add book/buy/ask for availability button
The total price of the selected options (quantity, variants, additional and shipping) is fetched and displayed on every option selection.
The button for booking/buying and asking for availability has been added. This button is hidden when user browses his own listing.
Add one loading indicator for all options instead of having loading indicator of each one of them.