fix: Fix the random loading issue
The problem is that listing details screen is getting into the loading state
unexpectedly when another screen is pushed above it. So it turns out that there
are two reasons for this problem:
1- Widgets that call media query to calculate width and height are getting
rebuilt in some cases (like when the soft keyboard appears) and this is causing
the screen to load and the state to bo lost.
To avoid this, the media query is initialized once at the beginning of the app.
2- Listening to http error state in listing details screen causes the screen
to rebuild when: an invalid booking date is picked (i.e. the listener gets
triggered) and another screen is pushed to the stack (above listing details screen).
To avoid this, the listener is moved to the date range picker widget.
Since this problem has been solved, report listing screen is now pushed above
listing details screen instead of replacing it.
Also, the select feature of riverpod has been used in different places to listen to
a specific state field instead of listening to the whole state object.
Lastly, some widgets in listing details screen have been simplified and some code
enhancements have been done.