fix: Fix an issue introduced in the previous commit
Also this is a fix for an issue introduced when the loading issue had been fixed.
The issue is that the variables used in login and register forms are not getting updated
when the user enters the data in the fields so the auth models stay empty and login or
register requests never get succeeded.