feat: Add edit listing functionality feat: Use "select" riverpod feature to listen to state feat: Add an HTML editor feat: Add routes observer
To add edit listing functionality, the following things have been implemented:
- Getting the listing for editing and then populating the fields with the fetched data.
- Initializing stock, price, min period and max period text fields.
- Initializing the photos (if any) of the listing to be edited.
- Initializing the availability status of the listing ("disable" status has been handled too).
- Refreshing the listing in listing details screen when the user edits the listing.
- Refreshing the listings in list and grid screens when the user edits the listing.
- Refreshing the listing in map screen when the user edits the listing.

The "select" feature of riverpod has been used to listen to a specific part of the state
(instead of listening to the whole state) so the number of rebuilds is reduced (read about select in
riverpod documentations for more info).

A new package has been used to add an html editor for editing the listing's description in step 4
of posting/editing a listing. After adding this package, the multiDex in Android had to be enabled.
IMPORTANT NOTE: the html editor is preventing the location picker (which is in the same screen) from
animating the map when the user picks a location.