feat: Add place order functionality [WIP]
This commit is related to the previous two commits and the following few commits will be related to
adding place order functionality.
In this commit, the required values are checked before navigating the user to place order screen.
For listings that are for rent: only the booking date is checked. For listings that are for sell:
shipping option and variations are checked.
Note: the number of files is big in this commit because an unsed import has been removed from many files.