feat: Add place order feature [WIP]
In this commit, summary tab view has been populated with place order summary data.
Also, a UI problem that happens when the text of the additional options (in listing details screen)
is long has been resolved.
The following problem [when the quantity of the selected additional option is one, the quantity received
in the pricing object of the place order details response is null] has been resolved too.